requirements.txt

Flask==2.3.2
flask-cors==4.0.0
opencv-python-headless==4.9.0.80
pillow==10.2.0
imagehash==4.3.1
scikit-image==0.22.0
numpy
